[ADD] ticket creation, and refactoring.
