+++ .gitignore
@@ -1,28 +1,6 @@
-# Compiled source
-*.class
-*.dll
-*.exe
-*.o
-*.so
-
-# Packages
-*.7z
-*.dmg
-*.gz
-*.iso
-*.jar
-*.rar
-*.tar
-*.zip
 
 # Logs and databases
 *.log
-*.sql
-*.sqlite
-
-# OS generated files
-.DS_Store
-Thumbs.db
 
 # Python
 __pycache__/
@@ -32,36 +10,7 @@ __pycache__/
 *.env
 .env/
 
-# Node.js
-node_modules/
-npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
-
-# Java
-*.war
-*.ear
-
-# Ruby
-*.gem
-*.rbc
-.bundle/
-vendor/bundle/
-
-# Go
-*.exe
-*.test
-*.out
-
-# Rust
-/target/
-
-# C/C++
-*.obj
-*.elf
-
 # IDEs and editors
-.idea/
 .vscode/
 *.swp
 *.swo
@@ -72,9 +21,8 @@ vendor/bundle/
 *.patch
 *.diff
 
-# Other
-*.bak
-*.tmp
-*.temp
-*.old
-*.backup
+# project specific
+data/
+docs/
+venv/
+.venv
