+++ src/agent/config/config.py
@@ -1,6 +1,11 @@
 """API configuration and constants"""
 
-BASE_URL = "https://portal.uat.karmayogibharat.net/api"
+import os
+from dotenv import load_dotenv
+
+load_dotenv()
+
+BASE_URL = os.getenv("BASE_URL")
 
 API_ENDPOINTS = {
     "CERTIFICATE": f"{BASE_URL}/course/batch/cert/v1/issue",
@@ -10,14 +15,16 @@ API_ENDPOINTS = {
     "USER_SEARCH": f"{BASE_URL}/private/user/v1/search" 
 }
 
-DEFAULT_HEADERS = {
-    "Accept": "application/json",
-    "Content-Type": "application/json"
+LLM_CONFIG = {
+    'temperature' : 1,
+    'top_p' : 0.9,
+    'top_k' : 40,
+    'max_output_tokens' : 4096,
+    'response_mime_type' : 'text/plain'
+
 }
 
 REQUEST_TIMEOUT = 60
 
-AUTH_CONFIG = {
-    "AUTH_URL": f"{BASE_URL}/auth/realms/sunbird/protocol/openid-connect/token",
-    "CLIENT_ID": "admin-cli"
-}
+TICKET_DIR = os.path.join(os.path.dirname(__file__), "../../../data/tickets")
+TICKET_FILE = os.path.join(TICKET_DIR, "tickets.json")
